Fix ResolvedCbBinding to use uintptr_t (CBHandle) for CB id
CBHandle is uintptr_t (64-bit), not uint32_t. Storing it truncated
caused UpdateDynamicCircularBufferAddress to receive garbage ids.
Also updates profiling test commentary to reflect that the fast path
is now active (create_descriptor not called on cache hits).

+++ tt_metal/api/tt-metalium/program_descriptors.hpp
@@ -216,8 +216,9 @@ struct ResolvedRtArgBinding {
 };
 
 // A CB dynamic-address binding resolved to a CB id for direct update.
+// cb_id is CBHandle (uintptr_t) — a 64-bit handle, not an index.
 struct ResolvedCbBinding {
-    uint32_t cb_id = 0;
+    uintptr_t cb_id = 0;
     Buffer* buffer = nullptr;
 };
